Replace `kodi_language` setting with code
Instead of making a query in a single language (which can be
misconfigured in many places), make queries in up to 3 different
languages.

payload['title'] is always looked up. It's in Kodi's language.

Make extra queries for a title [1]:

* in the movie's original language
* in the provider's language [2]

This logic is disabled when the title is forced - `{title:lang}`.

---
[1]: only when it differs from payload['title']
[2]: only when it's defined and if the movie has a title in it